CR - Make exception management more specific

+++ importer/lambda/cross_region_importer.py
@@ -130,30 +130,33 @@ def _delete_cross_stack_references(exports_to_remove, importer_context, table_in
                 Key={'CrossStackRefId': cross_stack_ref_id},
                 ConditionExpression=Attr('CrossStackRefId').eq(cross_stack_ref_id),
             )
-        except ClientError:
-            print(f'{cross_stack_ref_id} was not found, scanning to get the key name')
-            scan_response = cross_stack_ref_table.scan(
-                FilterExpression=
-                Key('ExportName').eq(export_name) &
-                Key('ImporterStackId').eq(importer_context.stack_id) &
-                Key('ImporterLabel').eq(label) &
-                Key('ImporterLogicalResourceId').eq(importer_context.logical_resource_id)
-            )
-            cross_stack_ref_ids = [
-                cross_ref_id['CrossStackRefId'] for cross_ref_id in scan_response['Items']
-                if cross_ref_id != cross_stack_ref_id
-            ]
-
-            while scan_response.get('LastEvaluatedKey'):
-                scan_response = cross_stack_ref_table.scan(ExclusiveStartKey=scan_response['LastEvaluatedKey'])
-                cross_stack_ref_ids.extend([cross_ref_id['CrossStackRefId'] for cross_ref_id in scan_response['Items']])
-
-            for ref_id in cross_stack_ref_ids:
-                print(f'Deleting {ref_id}')
-                cross_stack_ref_table.delete_item(
-                    Key={'CrossStackRefId': ref_id},
-                    ConditionExpression=Attr('CrossStackRefId').eq(ref_id),
+        except ClientError as e:
+            if 'The conditional request failed' in str(e):
+                print(f'{cross_stack_ref_id} was not found, scanning to get the key name')
+                scan_response = cross_stack_ref_table.scan(
+                    FilterExpression=
+                    Key('ExportName').eq(export_name) &
+                    Key('ImporterStackId').eq(importer_context.stack_id) &
+                    Key('ImporterLabel').eq(label) &
+                    Key('ImporterLogicalResourceId').eq(importer_context.logical_resource_id)
                 )
+                cross_stack_ref_ids = [
+                    cross_ref_id['CrossStackRefId'] for cross_ref_id in scan_response['Items']
+                    if cross_ref_id != cross_stack_ref_id
+                ]
+
+                while scan_response.get('LastEvaluatedKey'):  # This loop manage pagination of results
+                    scan_response = cross_stack_ref_table.scan(ExclusiveStartKey=scan_response['LastEvaluatedKey'])
+                    cross_stack_ref_ids.extend([cross_ref_id['CrossStackRefId'] for cross_ref_id in scan_response['Items']])
+
+                for ref_id in cross_stack_ref_ids:
+                    print(f'Deleting {ref_id}')
+                    cross_stack_ref_table.delete_item(
+                        Key={'CrossStackRefId': ref_id},
+                        ConditionExpression=Attr('CrossStackRefId').eq(ref_id),
+                    )
+            else:
+                raise
 
 
 def _get_cloudformation_exports(target_region):
